% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{add_predictions}
\alias{add_predictions}
\title{Add predictions to a dataframe. More passthrough than modelr::add_predictions, espcially needed for se.fit and re.form}
\usage{
add_predictions(df, model, se.fit = TRUE, ...)
}
\arguments{
\item{...}{}
}
\description{
Add predictions to a dataframe. More passthrough than modelr::add_predictions, espcially needed for se.fit and re.form
}
