% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_compare.R
\name{baseline_compare}
\alias{baseline_compare}
\title{Compare values to a baseline}
\usage{
baseline_compare(
  val_df,
  compare_col,
  base_lev,
  values_col,
  group_cols = NULL,
  comp_fun,
  ...,
  failmissing = TRUE,
  names_to = "name",
  values_to = "value",
  zero_adjust = 0,
  onlyzeros = FALSE
)
}
\arguments{
\item{val_df}{dataframe of values}

\item{compare_col}{column or columns to use for grouping- needed if the baseline is one level in this column. Otherwise ignored}

\item{base_lev}{value for baseline. Can be a level in \code{compare_col}, or a scalar, or a vector of \code{length(unique(val_df$compare_col))}. A dataframe would be useful but not yet implemented.}

\item{values_col}{column containing the values. Can be character or tidyselect or bare names, though the last are fragile.}

\item{group_cols}{character, columns to group the baselining by. Often spatial or env_obj or similar.}

\item{comp_fun}{function to use for comparison comparing the first two arguments (\code{difference} and \code{relative} included here as wrappers for \code{-} and \code{/}). Can be character, or list-format, (both safe) or bare name (brittle)}

\item{...}{additional arguments to \code{comp_fun}}

\item{failmissing}{logical, default TRUE. Use \code{tidyselect::any_of} or \code{tidyselect::all_of} when \code{values_col} or \code{compare_col} are character. see \code{\link[=selectcreator]{selectcreator()}}}

\item{names_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}

\item{values_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}

\item{zero_adjust}{numeric (default 0) or \code{"auto"}, adjustment to data to
avoid zeros by adding \code{zero_adjust} to \code{abs(data)}, e.g shifting all data
away from zero, either positively or negatively. Zeros themselves are
shifted up or down randomly. Used for avoiding x/0, NaN, and Inf when
relativiszing and taking logs, primarily. Auto shifts by
\code{0.1*min(abs(data[data != 0]))}.}

\item{onlyzeros}{logical, default \code{FALSE}. Should all values be adjusted away from zero (\code{TRUE}) or only adjust zero values (\code{FALSE})?}
}
\value{
a tibble matching \code{val_df} (or a long version thereof if \code{val_df} is wide), with an added column for the reference level (named \code{ref_values_col}) and a column of the compared values (named \code{comp_fun_values_col})
}
\description{
This is used to automate comparison to baselines of various sorts without needing to do all the data wrangling by hand.
}
