% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_distances.R
\name{calc_a}
\alias{calc_a}
\title{Find area of a circle-circle intersection given two radii and distance between centers}
\usage{
calc_a(d, radii, radius1, radius2)
}
\arguments{
\item{d}{distance between centers}

\item{radii}{length-2 vector of the radii of the circles. Cannot use this and radius1, radius2}

\item{radius1}{radius of circle 1. Cannot be used with radii}

\item{radius2}{radius of circle 2. Cannot be used with radii}
}
\description{
Implementation of \url{https://mathworld.wolfram.com/Circle-CircleIntersection.html}
}
