% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{cluster_estimates_fixed}
\alias{cluster_estimates_fixed}
\title{Extraction from only fixed models}
\usage{
cluster_estimates_fixed(model, alldata)
}
\arguments{
\item{alldata}{}
}
\description{
Extraction from only fixed models
}
