% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{cluster_estimates_rand}
\alias{cluster_estimates_rand}
\title{Extraction from re models}
\usage{
cluster_estimates_rand(model, alldata)
}
\arguments{
\item{alldata}{}
}
\description{
Extraction from re models
}
