% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_compare.R
\name{create_base}
\alias{create_base}
\title{Creates a baseline column in several different ways. Helper for \code{\link[=baseline_compare]{baseline_compare()}} but useful elsewhere too}
\usage{
create_base(
  val_df,
  compare_col,
  base_lev,
  values_col,
  group_cols = NULL,
  failmissing = TRUE,
  names_to,
  values_to
)
}
\arguments{
\item{val_df}{dataframe of values}

\item{compare_col}{column or columns to use for grouping- needed if the baseline is one level in this column. Otherwise ignored}

\item{base_lev}{value for baseline. Can be a level in \code{compare_col}, or a scalar, or a vector of \code{length(unique(val_df$compare_col))}. A dataframe would be useful but not yet implemented.}

\item{values_col}{column containing the values. Can be character or tidyselect or bare names, though the last are fragile.}

\item{group_cols}{character, columns to group the baselining by. Often spatial or env_obj or similar.}

\item{failmissing}{logical, default TRUE. Use \code{tidyselect::any_of} or \code{tidyselect::all_of} when \code{values_col} or \code{compare_col} are character. see \code{\link[=selectcreator]{selectcreator()}}}

\item{names_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}

\item{values_to}{character, as in \code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}, used for auto-pivotting}
}
\value{
a tibble matching \code{val_df} (or a long version thereof if \code{val_df} is wide), with an added column for the reference level (named \code{ref_values_col})
}
\description{
Creates a baseline column in several different ways. Helper for \code{\link[=baseline_compare]{baseline_compare()}} but useful elsewhere too
}
