% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{extract_cluster_terms}
\alias{extract_cluster_terms}
\title{Extract cluster coefficients from full models and adjust by slope/intercept if needed.
Tends to be the same as the predictions from the full model, but with different se.}
\usage{
extract_cluster_terms(model, alldata)
}
\arguments{
\item{alldata}{}
}
\description{
Extract cluster coefficients from full models and adjust by slope/intercept if needed.
Tends to be the same as the predictions from the full model, but with different se.
}
