% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{extract_glmmTMB_ests}
\alias{extract_glmmTMB_ests}
\title{Extract the term estimates from glmmTMB}
\usage{
extract_glmmTMB_ests(model, mname)
}
\arguments{
\item{model}{}
}
\description{
Extract the term estimates from glmmTMB
}
