% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_distances.R
\name{find_d}
\alias{find_d}
\title{Find the distance between centers given radii and overlap area}
\usage{
find_d(area, radii, radius1, radius2)
}
\arguments{
\item{area}{area of overlap of the circles}

\item{radii}{length-2 vector of the radii of the circles. Cannot use this and radius1, radius2}

\item{radius1}{radius of circle 1. Cannot be used with radii}

\item{radius2}{radius of circle 2. Cannot be used with radii}
}
\description{
Find the distance between centers given radii and overlap area
}
