% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_pages.R
\name{find_missing_pages}
\alias{find_missing_pages}
\title{Find quarto files not in _quarto}
\usage{
find_missing_pages(files = NULL, pattern = "_quarto.*.yml")
}
\arguments{
\item{pattern}{}
}
\description{
Find quarto files not in _quarto
}
