% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinAR.R
\name{fit_nbinomAR}
\alias{fit_nbinomAR}
\title{Estimate parameters of an autocorrelated negative binomial}
\usage{
fit_nbinomAR(X)
}
\arguments{
\item{X}{}
}
\value{
a tibble with the fitted parameter values for size, mu, and rho
}
\description{
Wraps \code{\link[stats:acf]{stats::acf()}} and \code{\link[fitdistrplus:fitdistr]{fitdistrplus::fitdistr()}}
}
