% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gaussian_mm_data.R
\name{get_clusterobs}
\alias{get_clusterobs}
\title{Get the observations in each cluster}
\usage{
get_clusterobs(
  n_clusters,
  N,
  xvals,
  cluster_N = "fixed",
  nobs_mean = "fixed",
  force_nclusters = TRUE,
  force_N = TRUE,
  nbsize,
  uneven_with_x = FALSE
)
}
\arguments{
\item{n_clusters}{Number of clusters (levels of random factor)}

\item{N}{total number of observations}

\item{cluster_N}{character, 'fixed', 'uneven', with fixed yielding
N/n_clusters per cluster, uneven being a random number of obs in each
cluster chosen by rnbinom. Can be vector if using subclusters and want
different values at different levels}

\item{nobs_mean}{character, 'fixed', 'x'. If \code{cluster_N = 'uneven'}, this
controls whether the \code{mu} term in \code{\link[stats:NegBinomial]{stats::rnbinom()}} is the same for all
clusters (n_clusters/N) or varies (with x, but that may be shuffled
according to \code{uneven_with_x})}

\item{force_nclusters}{logical- force the number of realised clusters to
equal n_clusters. Done by adding 1 to any clusters that randomly end up
with 0. Can be vector if using subclusters and want different values at
different levels}

\item{force_N}{logical- force the total realised obs to equal N. Done by
sampling from the realised obs in clusters given its probabilities until we
get N.}

\item{nbsize}{\code{size} argument to \code{\link[stats:NegBinomial]{stats::rnbinom()}}}

\item{uneven_with_x}{logical- keep any uneven rnbinom \code{mu} correlated with x
(TRUE) or shuffle relative to x (FALSE)}
}
\description{
Get the observations in each cluster
}
