% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_multi.R
\name{ggsave_multi}
\alias{ggsave_multi}
\title{Vectorised version of ggsave}
\usage{
ggsave_multi(filename, plot = last_plot(), device, ...)
}
\arguments{
\item{filename}{As in \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, but without extension, and no extension guessing.}

\item{plot}{As in \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}

\item{device}{As in \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, but must be character and can be a character vector}

\item{...}{All other arguments to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\value{
invisible vector of filenames
}
\description{
Passes all arguments to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} but is vectorised over \code{device} and \code{filename} should not have an extension
}
