% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_yaml.R
\name{make_multi_yaml}
\alias{make_multi_yaml}
\title{Sets up a main and subsidiary yaml. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}}
\usage{
make_multi_yaml(
  yamdir = "_yml",
  copy_orig = TRUE,
  make_simple = TRUE,
  leave_orig = TRUE
)
}
\arguments{
\item{leave_orig}{}
}
\description{
Sets up a main and subsidiary yaml. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}
}
