% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pal.R
\name{make_pal}
\alias{make_pal}
\title{Function to ensure colors match across plots. Developed in various places, currently mainly owned by \href{github.com/MDBAuth/HydroBOT}{HydroBOT}.}
\usage{
make_pal(
  levels,
  palette,
  refvals = NULL,
  refcols = NULL,
  includeRef = FALSE,
  returnUnref = FALSE,
  direction = 1
)
}
\arguments{
\item{levels}{character vector of levels to use (should be all possible for the variable)}

\item{palette}{character vector of palette name from {paletteer}}

\item{refvals}{character vector of levels we want to set manually (typically as references)}

\item{refcols}{character vector (names or hex) of colors to set manually in order matching \code{refvals}}

\item{includeRef}{logical, defailt FALSE. Should the colors be chosen including (TRUE) or not including (FALSE, the default) the levels in \code{refvals}. Typically \code{TRUE} if we sometimes want to accentuate a level, and sometimes not.}

\item{returnUnref}{logical, default FALSE. Should we return the base palette with the refvals matched to \code{palette} in addition to the version where they match \code{refcols} (TRUE). Only works if \code{includeRef = TRUE}. If \code{TRUE}, returns a list with the base palette and the palette matching \code{refcols}}

\item{direction}{1 (default) or -1 (reverse the palette), as in \code{\link[paletteer:paletteer_c]{paletteer::paletteer_c()}} and \code{\link[paletteer:paletteer_d]{paletteer::paletteer_d()}}}
}
\value{
typically a named character vector of colors, unless \code{includeRef = TRUE} and \code{returnUnref = TRUE}. In that case, a list of two named color vectors.
}
\description{
Function to ensure colors match across plots. Developed in various places, currently mainly owned by \href{github.com/MDBAuth/HydroBOT}{HydroBOT}.
}
