% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_yaml.R
\name{make_simple_yaml}
\alias{make_simple_yaml}
\title{Rebuilds _quarto.yml to be a single file. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}}
\usage{
make_simple_yaml(proj_yaml_file = NULL, yamdir = getwd(), simple_file = NULL)
}
\arguments{
\item{simple_file}{}
}
\description{
Rebuilds _quarto.yml to be a single file. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}
}
