% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_simpleyml.R
\name{make_simpleyml}
\alias{make_simpleyml}
\title{Quick helper to create a quarto profile to only render the active file}
\usage{
make_simpleyml(renderfile = "auto")
}
\arguments{
\item{renderfile}{file to render, default 'auto' uses the active file in Rstudio}
}
\description{
Quick helper to create a quarto profile to only render the active file
}
