% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{naive_clusters}
\alias{naive_clusters}
\title{Just get the mean and sd of each cluster}
\usage{
naive_clusters(alldata)
}
\arguments{
\item{alldata}{}
}
\description{
Just get the mean and sd of each cluster
}
