% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinAR.R
\name{nbin_chi}
\alias{nbin_chi}
\title{Chi-square for nbin}
\usage{
nbin_chi(distdf, grouper)
}
\arguments{
\item{distdf}{return from \code{\link[=nbin_emp_pmf]{nbin_emp_pmf()}}}
}
\description{
Chi-square for nbin
}
