% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinAR.R
\name{nbin_emp_pmf}
\alias{nbin_emp_pmf}
\title{Get empirical and negative binomial distributions}
\usage{
nbin_emp_pmf(X, size, prob, mu)
}
\arguments{
\item{X}{vector of values}

\item{size}{target for number of successful trials, or dispersion
    parameter (the shape parameter of the gamma mixing distribution).
    Must be strictly positive, need not be integer.}

\item{prob}{probability of success in each trial. \code{0 < prob <= 1}.}

\item{mu}{alternative parametrization via mean: see \sQuote{Details}.}
}
\description{
Particularly useful for checking fits look good
}
