% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbinAR.R
\name{nbin_gg}
\alias{nbin_gg}
\title{Diagnostic plot of empirical and theoretical negative binomial}
\usage{
nbin_gg(distdf)
}
\arguments{
\item{distdf}{return from \code{\link[=nbin_emp_pmf]{nbin_emp_pmf()}}}
}
\description{
Diagnostic plot of empirical and theoretical negative binomial
}
