% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn_distances.R
\name{opt_d}
\alias{opt_d}
\title{Function to optimize to find distance between circle centers}
\usage{
opt_d(d, area, radii, radius1, radius2)
}
\arguments{
\item{d}{distance between centers}

\item{area}{area of overlap of the circles}

\item{radii}{length-2 vector of the radii of the circles. Cannot use this and radius1, radius2}

\item{radius1}{radius of circle 1. Cannot be used with radii}

\item{radius2}{radius of circle 2. Cannot be used with radii}
}
\description{
Function to optimize to find distance between circle centers
}
