% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{predict_at_clusters}
\alias{predict_at_clusters}
\title{Get full-model predictions for each cluster (and x value if present)}
\usage{
predict_at_clusters(model, alldata)
}
\arguments{
\item{alldata}{}
}
\description{
Get full-model predictions for each cluster (and x value if present)
}
