% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_models.R
\name{predict_fit}
\alias{predict_fit}
\title{Get model predictions across an x-range at the mean random
This is a bit silly with how close it is to the above, but whatever.}
\usage{
predict_fit(model, xsteps)
}
\arguments{
\item{alldata}{}
}
\description{
Get model predictions across an x-range at the mean random
This is a bit silly with how close it is to the above, but whatever.
}
