% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2hex.R
\name{rgb2hex}
\alias{rgb2hex}
\title{Convert R, G, B columns to hex values}
\usage{
rgb2hex(R, G, B, maxval = 255)
}
\arguments{
\item{R}{}

\item{G}{}

\item{B}{}

\item{maxval}{}
}
\description{
Convert R, G, B columns to hex values
}
