% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_gaussian_mm_data.R
\name{simulate_gaussian_mm}
\alias{simulate_gaussian_mm}
\title{Simulate data with random effects structure.}
\usage{
simulate_gaussian_mm(
  n_clusters,
  n_subclusters = NULL,
  N,
  intercept,
  slope,
  obs_sigma,
  sd_rand_intercept,
  sd_rand_slope,
  rand_si_cor,
  obs_x_sd,
  cluster_N = "fixed",
  nobs_mean = "fixed",
  force_nclusters = TRUE,
  force_nsubclusters = TRUE,
  force_N = TRUE,
  cluster_x,
  nbsize,
  uneven_with_x = FALSE,
  x_is_density = FALSE
)
}
\arguments{
\item{n_clusters}{Number of clusters (levels of random factor)}

\item{n_subclusters}{Number of subclusters (total). Can be a vector to have
multiple levels of nesting}

\item{N}{total number of observations}

\item{intercept}{True intercept of the relationship between x and y}

\item{slope}{True slope of the x-y relationship}

\item{obs_sigma}{Observation-level sd, i.e. variance relative to the cluster
mean (residual variance)}

\item{sd_rand_intercept}{Standard deviation of the random intercept, i.e.
constant variance of the random levels around the line. If using
subclusters, should be a vector with a value for each scale, starting with
cluster, then subcluster 1, 2, ...}

\item{sd_rand_slope}{Standard deviation of the random slope, i.e. how much
does the slope of the x-y relationship vary between random levels. If using
subclusters, should be a vector with a value for each scale, starting with
cluster, then subcluster 1, 2, ...}

\item{rand_si_cor}{Correlation between random slope and intercept.If using
subclusters, should be a vector with a value for each scale, starting with
cluster, then subcluster 1, 2, ...}

\item{obs_x_sd}{Standard deviation of observational x-variation around
cluster means. If 0, all obs in a cluster have the same value. If cluster_x
is length-1, then this just yields a common distribution for all obs.
\emph{Note}, this is not set up to have differences in the x of different
subclusters/nestings- the x determination is done for observations within
the largest cluster.}

\item{cluster_N}{character, 'fixed', 'uneven', with fixed yielding
N/n_clusters per cluster, uneven being a random number of obs in each
cluster chosen by rnbinom. Can be vector if using subclusters and want
different values at different levels}

\item{nobs_mean}{character, 'fixed', 'x'. If \code{cluster_N = 'uneven'}, this
controls whether the \code{mu} term in \code{\link[stats:NegBinomial]{stats::rnbinom()}} is the same for all
clusters (n_clusters/N) or varies (with x, but that may be shuffled
according to \code{uneven_with_x})}

\item{force_nclusters}{logical- force the number of realised clusters to
equal n_clusters. Done by adding 1 to any clusters that randomly end up
with 0. Can be vector if using subclusters and want different values at
different levels}

\item{force_N}{logical- force the total realised obs to equal N. Done by
sampling from the realised obs in clusters given its probabilities until we
get N.}

\item{cluster_x}{How to handle x-values of the cluster. If numeric and length
n_clusters, it gives the x of each cluster. If a function, e.g. \(x)
runif(x, min = -1, max = 2) or \(x) rnorm(x), it chooses the cluster-mean
x's from those distributions. In the special case of a length-1 numeric, it
gives a common value to all clusters, which is particularly useful if you
want the observations to vary along x but not the clusters themselves (as
the original function did).}

\item{nbsize}{\code{size} argument to \code{\link[stats:NegBinomial]{stats::rnbinom()}}}

\item{uneven_with_x}{logical- keep any uneven rnbinom \code{mu} correlated with x
(TRUE) or shuffle relative to x (FALSE)}

\item{x_is_density}{logical- re-calculate x directly from the number of obs
in the cluster; e.g. x \emph{is} density.}
}
\value{
a dataframe of simulated data
}
\description{
This is only very slightly modified from Clark (2019, May 14), which is an
\emph{excellent} visual treatment of shrinkage in mixed models. Michael Clark:
Shrinkage in Mixed Effects Models. Retrieved from
https://m-clark.github.io/posts/2019-05-14-shrinkage-in-mixed-models/
Modifications include much more control over how x works and the form of
unbalancing, along with renaming and shifting some control to the user. The
choice of x structure is important here, and reflects sampling strategies and
experimental design. I've added a lot of ability to control it, though we may
still need more. This has one level of random factor, 'cluster', with some
number of observations within it.
}
