% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_clean}
\alias{theme_clean}
\title{A standard ggplot theme for HydroBOT}
\usage{
theme_clean(base_size = 10, ...)
}
\arguments{
\item{base_size}{set the base_size as in other \code{ggplot} themes}

\item{...}{other arguments to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
a ggplot theme
}
\description{
A standard ggplot theme for HydroBOT
}
\examples{
ggplot2::ggplot(iris,
  ggplot2::aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
  ggplot2::geom_point() +
  theme_clean()
}
