% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_yaml.R
\name{use_quarto_yaml}
\alias{use_quarto_yaml}
\title{Switches among several _quarto.yml. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}}
\usage{
use_quarto_yaml(yamfile = "project", yamdir = "_yml")
}
\arguments{
\item{yamdir}{}
}
\description{
Switches among several _quarto.yml. \emph{ALMOST ALWAYS BETTER TO USE \code{\link[=make_simpleyml]{make_simpleyml()}} with a profile}
}
