% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipdownload.R
\name{zipdownload}
\alias{zipdownload}
\title{Helper to download and unzip and throw out extras}
\usage{
zipdownload(dirname, datadir, sourceurl, existing_dirs = list.files(datadir))
}
\arguments{
\item{existing_dirs}{}
}
\description{
Helper to download and unzip and throw out extras
}
